kafka: add an ignored sync_send() option to kafka for compatibility with the SCL
Sending messages to kafka synchronously does not really make sense, as it
will be very slow, so I don't expect anyone to actually want to use it.

This patch adds an ignored sync-send() option, so we accept this option in
existing configurations.  At the same time, we display a warning and ignore it.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>