kafka: add support for multiple options formats
This patch implements a new-style for options, namely

"name" => "value"

At the same time it allows Java style options to make it compatible with
Java/Python which uses a list of pairs, and it retains the original
name("value") format as well.

The plan is to allow any other options() like construct to reuse this
syntax in the future.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>