http: abort the current batch in case we are going down for a reload
In case of a reload (vs. a shutdown) we can easily drop the current batch
as that will wait for us nicely in the queue when we start back up.

At shutdown we are not so fortunate, as we will only be able to retain
the messages if the user has a disk buffer.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>