kafka: make the kafka implementations trivial to flip
The rdkafka based destination is renamed to kafka-c(), and the
kafka SCL is modified so that it uses the java based one by default.

If you want to flip to the rdkafka based one, just:

1) @define kafka-implementation to kafka-c
2) replace kafka with kafka-c() in your config

In the future, we will be deprecating kafka-java() and possibly the entire
feature flip functionality, and at that point kafka-c() will be
aliased to kafka().

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>