http: allow URLs to be specified by a space separated string
It is pretty common that we get a list of things via a string argument. For
instance, for Java/Python based destinations that's the only way to
pass a list. An example is the java based kafka driver, where
kafka-bootstrap-servers() is one such attribute.

I recently encountered a case, where someone wanted to use a similar format
to HTTP and was surprised it didn't support this case.

This patch changes the url() parameter, it now supports both cases:

  * url("server1", "server2", "server3"); #commas are optional
  * url("server1 server2 server3");

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>