Add support for usec precision when parsing time
Python added a '%f' field ("Microsecond as a decimal number") in it's
strptime() implementation to parse times with subsecond precision.

Borrow this idea and implement it in syslog-ng using the same field
descriptor.

Eat all digitis when reading microseconds: if the field has a nanosecond
precision, the value will be truncated at the microsecond and the parser
will be able to process the information following the sub-second
information.