filters: hide FilterRE struct
Our convention is that various setters use the base class as their
first argument and constructors also return that type. This convention
makes it possible to hide the actual class in the implementation file.

This convention was not followed in the filter-re.c module and this
patch aims at improving the situation.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>