filters: add support for template() option in match()
This patch adds support for a template() option in match() thereby
allowing to match against constructed values.

In case the template is trivial (e.g. only literal text
or a single name-value pair reference), performance is the same as
using value(), so using template() instead of value() is the same performance
wise but more consistent syntax wise.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>