lib/timeutils: Reset the timezone on config reload event.
Currently, the behaviour of syslog-ng regarding timezone changes handling
depends on SYSLOG_NG_HAVE_LOCALTIME_R configuration parameter. If
SYSLOG_NG_HAVE_LOCALTIME_R is set to '1', the timezone is not updated
in run time and vise versa. However, timezone setting is considered as
a costly operation, so it can be not desired to perform it
frequently. One of the wice options is setting the timezone on-demand.
The code sets timezone on config reload event.

signed-off-by: Oleksii Hamov <oleksii.hamov@ring.com>