cfg-lex: track escaped quote characters within strings that are within blocks properly
A string such as "foo\"" will contain an embedded quote character. When
such thing is encountered within a block, we need to track the
string boundary and include the string in its literal form in the
block value.

Previously this was not working properly because of how flex treats
conflicting rules. That is:

  1) the longest match is found
  2) if multiple rules have the same match length, the one first in the lexer wins

Previously we had a wildcard rule for continuation characters, that
simply removed the "\\\n" sequence while updating the the location. This
is wrong for blocks, where we want to retain everything literally.

This change will effect that change in the lexer and adds a couple
of testcases to make sure it does not break in the future.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>