test_userdb: Use the sys group and fix hard-coded GID
The FreeBSD operating system does not have a "root" group, and GID 0 is
used for the "wheel" group.  POSIX does not allocate GID to specific
groups, and does not provide a list of groups that should exist, so
instead of assuming that the "root" group has GID 0, lookup the GID of
group "sys" and use this for the check (a "sys" group exist on all
supported plateforms and was available in research UNIX V7 (1972) which
makes it a good candidate for being found virtually everywhere).

This fix the test when running on FreeBSD.

Signed-off-by: Romain Tartière <romain@blogreen.org>