python-functional: Fix hostname matching on FreeBSD
Linux systems are usually configured with the hostname set to the short
name of the system.  The fully qualified domain name is computed by
using this configured short hostname and querying the domain name system
(or the /etc/hosts file).

FreeBSD (and probably all *BSD) are configured with the fully qualified
domain name, and when the short name of the system is requested, it is
built by returning the fully qualified domain name up to the first dot
it contains.

A consequence is that gethostname(3) returns a short hostname on Linux
and a fully qualified domain name on other systems.

Python's socket.gethostname() relies on gethostname(3) and return a short
hostname on Linux and a fully qualified domain name on other systems.

Because the test suite is expected to find a short hostname, add a bunch
of functions:
  - get_hostname() to query the hostname as seen by gethostname(3);
  - get_short_hostname() to query the short hostname;
  - get_fqdn() to query the fully qualified domain name.

This fix these tests when running them on FreeBSD.

Signed-off-by: Romain Tartière <romain@blogreen.org>