cmake: introduce check target
This patch introduces a separate "check" target for CMake generated
makefiles as described here:

https://gitlab.kitware.com/cmake/community/wikis/doc/tutorials/EmulateMakeCheck

Thisx allows us to only build test programs in case we are running tests,
and to couple rebuilding of the test programs with running them.

This also allows us to reference the unit test phase as a DEPENDS in
add_custom_target(), as the "test" target is special, whereas "check" is a
normal target.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>