afsocket-source: adjust dynamic-window-size()
Let's imagine, that the user sets dynamic-window-size(200) and
max-connections(101).

At some point in time, there are 100 connections,
so they all get 2 window each.
Then another connection is established, and as we can't give every
connection 2 window each, they all get 1.
So with that one new connection we dropped from using 200 windows
to using only 101, this is nearly 50% drop.

If we set the dynamic-window-size() to at least a 10 or so-ish multiple
of max-connections() we drastically drop the impact of the problem.
This suggestion is raised as warning.

We can tweak this even better, with syncing these 2 options,
by changing dynamic-window-size() to be a multiple of max-connections().
We adjust dynamic-window-size, and raise a warning, that we changed it.

Signed-off-by: Attila Szakacs <attila.szakacs@balabit.com>