travis: eliminating if related to --keep-going
The idea was that we call make with --keep-going. But with that, we
can lose track of the last error. Hence the make command is restarted
without keep-going, making sure the we can see some error at the
bottom of the log.

I can do both in the first line:
```make --keep-going -j $(nproc) || make V=1 install;```

With that change, the if is not needed anymore

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>