build: remove -Wcast-align
This warning was never fixed, but gcc is take the ABI into
account when reporting alignment issue. In case of x86 those are
not actual issues, thus gcc silence.

On the other hand clang reports it despite of arch.

Also additionally the most issue is related to *sockaddr* casting,
and some magic inside syslog-ng related to nv-pairs (both are probably okay on arm as well).

Signed-off-by: Kokan <kokaipeter@gmail.com>