travis: remove --keep-going from distcheck
Similarly to the normal build, we add --keep-going to get as many
errors as possible. Because of this, errors are somewhere in the
middle of the log. So we restart the target so that we could see an
error in the end of the log.

The restart logic unfortunately does not work for distcheck target,
because after a failed distcheck attempt, make distcheck will quickly
exit because it expects a _build directory not available.

Return code 42 was used to to make distcheck error and propagation
error distinguishable. Only in case of 42 (distcheck error) we should
have re-executed distcheck. After --keep-going removal, the check for
42 is not needed anymore either.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>