build: remove -Wsuggest-attribute=noreturn
First of all the `-Wsuggest-attribute=noreturn` is only supported by `gcc`,
a better option would be `-Wmissing-noreturn` which is supported by `gcc` and
`clang` as well.

As the current standard the project follows is C99, there is no standard way,
to conform with this warning (only gcc-ism).

C11 provides `_Noreturn` keyword to save the day.

I do not consider this warning that that important, as there is warning
for function with non-void functions, this only handles methods
that does not return (terminates the executable).

Additionally the current flex 2.5.35 (used in travis) generates a code
that does not conform with `-Wmissing-noreturn` (does not use the needed
attribute). Newer version fixed this behaviore (2.6).

I would only enable this warning when C11 is the used standard,
by that time I hope flex newer version is also updated most of the places.

Signed-off-by: Kokan <kokaipeter@gmail.com>