timestamp: add fix-timezone() rewrite operation
The new fix-timezone() rewrite operation allows the user to fix up the timezone
value of a message assuming it was improperly parsed. The new timezone can be passed
as a timezone name or can even be a template string.

Examples:

Set the "stamp" timestamp of the message to EST or EDT, assuming it was
originally incorrectly recognized.

	rewrite { fix-time-zone("EST5EDT" time-stamp(stamp)); };

Use a name-value pair to store the appropriate timezone that can be set by something else:

	rewrite { fix-time-zone("${timezone}"); };

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>