lib/filter: fix standalone filter grammar parsing 
When specific parsers are called from the global config, they automatically
inherit all the global keywords. They only extend the list with their own
unique keywords. (static CfgLexerKeyword bar[] =...)  

Unlike other parsers, the filter parser is used as a standalone parser too.
(for example: grep/if templates, pdbtool) In such cases, the filter parser
only recognize keywords, which are declared in filter-expr-parser.c

Why this is a problem? The “filter” keyword has a double meaning. One is to
declare named filters at the global config level. The other one is to call
a previously declared filter (filter call). Because the "filter" keyword is
declared globally, a stand alone filter parser can not recognize it, and
simply handled it as an “LL_IDENTIFIER”, which results in parsing errors.

The solution is to declare the missing keyword in filter parser. This will
implicate that the keyword will be redundant when we are using the parser
normally, but fortunately the lexer handles this situation gracefully.

Signed-off-by: Kokan <kokaipeter@gmail.com>
Signed-off-by: Laszlo Szemere <laszlo.szemere@balabit.com>