tests/copyright/check.sh: implement ordering between licensing classes
Previously the check.sh script insisted that all matching license types
be associated with a file, collecting them in a comma separated list.

This patch on the other hand introduces ordering between the license
types instead, so if a file matches one license type the rest of them
will not be tried.

The implementation prepends the sequence number of the license type in
tests/copyright/policy to the name of the file in the working directory.
Then the matching would go through those files in this order and the
first one would win.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>