dbparser: advance()+flush() separately
_advance_time_based_on_message() is responsible for updating the time
based on the incoming message and for triggering pattern_db_expire_entry()
on expired contexts.

pattern_db_expire_entry() then executes the actions of the expired
context. In order to do this, it needs a PDBProcessParams instance
as an output object, so it can, for example, buffer emitted messages.

An expired context can change the fields of `process_params`, therefore
_advance_time_based_on_message() should not be called in the middle
of another rule execution, because it can lead to crashes.

To avoid such scenarios, _pattern_db_advance_time_and_flush_expired() has
been added, which triggers the actions of the expired contexts and flushes
the output separately.

Signed-off-by: László Várady <laszlo.varady@balabit.com>