python: initialize Python "program name" with Py_SetProgramName()
"The Py_SetProgramName() function should be called before Py_Initialize()
to inform the interpreter about paths to Python run-time libraries."

Setting it before Py_Initialize() is useful when syslog-ng is bundled
with its own Python interpreter, so it finds the bundled libraries
automatically. It does not cause problems when system Python is used,
because it falls back to system search paths.

Signed-off-by: László Várady <laszlo.varady@balabit.com>