light: network-ops: fix get_fqdn() (and get_short_hostname())
socket.getfqdn():
  Return a fully qualified domain name for name. To find the fully
  qualified name, the hostname returned by gethostbyaddr() is checked,
  followed by aliases for the host, if available. The first name which
  includes a period is selected. In case no fully qualified domain name
  is available, the hostname as returned by gethostname() is returned.

Unfortunately, this is not entirely true when a local name is resolved.

For example:
gethostname() == travis-job-1234

/etc/hosts:
127.0.1.1 travis-job-1234
127.0.0.1 localhost travis-job-1234

In this case, getfqdn(gethostname()) returns localhost, because it starts
to use (prefers) 127.0.0.1 instead of 127.0.1.1.

This is not an incorrect behavior, but Light should resolve names the same
way as syslog-ng does.

Signed-off-by: László Várady <laszlo.varady@balabit.com>