timeutils: fix cache cleanup upon reload
Whenever the configuration is reloaded we need to prepare the possibility
that our timezone is changed. In that case localtime/gmtime/mktime caches
need to be cleared.

The existing functionality merged recently was only doing this in the main
thread and not in worker threads. In most cases this was ok, as worker
threads are terminated across reloads, but any other threads may survive and
may also use the cached time functionality.

This patch introduces an atomic get in the fastpath, but this shouldn't
be that much of a problem.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>