persist-state: fix foreach
a) Fixed a mapping issue around PersistValueHandler.
b) It is possible that the PersistEntry is freed by the callback...
this means that the unmap is called with an invalid value. The free is
not explicitly called by the user, it can be a result of a
`persist_state_alloc_string()` for an existing key.
So we need to save the original offset that can be safely use after the
callback.

Signed-off-by: Laszlo Budai <laszlo.budai@outlook.com>