nvregistry: name_map must own a copy of name
The nvregistry consists of an array of NVHandleDesc(riptors) and a name_map,
that helps lookup the NVHandleDesc via its name property. In the current code
the ownership is at least confusing:

```
NVHandleDesc stored;
...
stored.name = g_strdup(name);
nvhandle_desc_array_append(self->names, &stored);
g_hash_table_insert(self->name_map, stored.name, ...);
```

From this I would assume that the *stored.name* owns the new pointer, and the
*self->name_map* has a view for that (reference). But when it comes down to
freeing the memory, it is the *self->name_map* that frees the pointer which
should fall to the owner. While the *self->name_map* could hold names that is not
owned by any NVHandleDesc (alias names). The map itself has to free *some* keys,
the proposal is to copy the name for each map entry. The performance penalty of
such thing should not be high, as it is only done for each new name which should
converge on the longer run.

Note: this introduces a leak, which is going to be fixed in the next patch.

Signed-off-by: Kokan <kokaipeter@gmail.com>