filter-expressions: add missing filter_expr_init calls
If a module uses filters for a subtask (ie.: rewrite rules
as an optional condition), than that filter is parsed and
created locally in that module.

Unfortunately we forgot to initialize those new filters in
some cases.

The problem could remain hidden, because in such cases, we
only use a very simple expression, and those simple filters
could work without initialization. (note: It is not necessary
for a filter to implement the init function.)

Signed-off-by: Laszlo Szemere <laszlo.szemere@oneidentity.com>