cfg-parser: add cfg_run_parser_with_main_context
While we are calling the parser of a module with cfg_run_parser
- during the processing of a syslog-ng config file - we pass the
globally used lexer (with all of it's keywords) to it.

There are keywords which have both meaning in the global context
and inside their module. (They don't have to be the same, they are
just formally identical.) i.e.: the "filter" keyword declares a
filter block in the global context and refers to a "filter call"
inside the filter module.

To avoid keyword duplications, we do not list those keywords again
inside the modules, we rather rely on the fact, that they are
already declared in the main parser.

The problem is, that if we use a parser outside of the context of
the main parser/config (i.e.: filters can be used as a condition
statement inside rewrite rules.) than the lexer will no longer
recognise the necessary keywords and handles them as a simple
identifier, which leads to parsing errors.

Signed-off-by: Laszlo Szemere <laszlo.szemere@oneidentity.com>