geoip2: test use maxmindb enterprise databse
Important, this also workarounds a memory leak found in the libmaxminddb, the
leak was fixed in upstream(https://github.com/maxmind/libmaxminddb/pull/180) like
a year ago, and no official release has been made since. As the leak is fixed,
but it blocks this project sanitizer progress, this change replaces the database
with a version that does not trigger the leak.

Signed-off-by: Kokan <kokaipeter@gmail.com>