java-modules: improve "make clean" logic
make distcheck requires us to remove the .gradle cache upon "make clean",
but sometimes I want to retain that when working interactively.  (e.g.  I
issue make clean on my own to trigger a rebuild, but then I lose Java jar
files as well).

This patch will allow one to override the cache directory with an
environment variable, in which case it is not removed.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>