cfg: introduce subordinate configurations
A subordinate config is one that relates to a "master" one, but is otherwise
independent.  This can be used to use secondary configs as used by
add-contextual-data().  I think this has been a horrible idea.

Currently a subordinate config is one that shares the set of available
modules with a master configuration.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>