diskq: rename `qdisk_get_size` to `qdisk_get_maximum_size`
The get_size() functions name is misleading, it suggests that it's the actual
size of the QDisk. In real it is the config option `disk_buf_size`.

It's not the first time I read the code and it misleads me, but I hope it
was the last time it tricked me.

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>
Signed-off-by: Kokan <kokaipeter@gmail.com>