diskq: create abstraction for record_len
record_length is a semi-hidden element in the disk-queue file.
Each serialized LogMessage manifested in a string, which length is stored in the disk-queue,
so we can read one LogMessage at a time.
(We know where does one message end and another one start)

Throughout the code the sizeof(guint32) or a sizeof(n) appers when the record_length
would be more meaningful.

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>
Signed-off-by: Kokan <kokaipeter@gmail.com>