logsource: add explicit (un)initialized state to WindowSizeCounter
Fixes: #2893

On 32 bit systems (or non-64 bit systems), syslog-ng could abort during
shutdown.

What was the reason of the abort?
a) in `log_source_set_options` where we set the initial window size
conditionally, the condition was false thus the `full_window_size`
remained 0
b) when `log_source_free` is called during shutdown,
 * `_release_dynamic_window` called unconditionally and
 *  a dynamic_part is calculated as full_window_size(=0) - init_window_size(=default 100),
 so dynamic_part = -100
 * window_size is decremented by dynamic_part(-100) and the
 `window_size_counter_sub` asserts on old_value >= value, and this
 assert failed, so syslog-ng aborted

So the questions are
1) why we did not set initial window size?
2) why old_value was not greater than value?

Answers:
1) the value we get from `window_size_counter_get` is the masked
value... on 64 bit systems this value is a 63 bits of `1` and it is compared to
a 32 bits of `1` but the 63 bits are truncated to 32 thanks to an explicit cast
And what if we are on a 32 bits system?
Well... the sizeof(gsize) is 4 , sizeof(gint) is also 4 on these
systems. This means that the `window_size_counter_get` returns 31 bits of
`-1`, and it is compared to 32 bits of `1` : they are obviously not
equals -> we won't set full_window_size

2) old_value is a -1, which is masked, so the actual old value is 2^31-1, while new value is a
-100, which is (2^32-100), so on a 32 bits system 31 bit negative value is
compared to a  to 32 bits negative value...

Proposed solution:
 * add a initialized state to LogSource: this is checked/(set to TRUE) only in
 `log_source_set_options`, and set to FALSE only in `log_source_init_instance`

Signed-off-by: Laszlo Budai <laszlo.budai@outlook.com>