logmsg: invalidate LEGACY_MSGHDR when PID or PROGRAM is unset
log_msg_set_value() contains the same invalidation logic for the
legacy message header. The same has to be done for unset() in order
to be consistent.

For example, the legacy message header is used by LogWriter if no
template is set, and unsetting the PID will not remove the PID field
from the output:

> rewrite { unset(value("PID")); };
> Aug  2 11:11:11 test test[2222]: test

> rewrite { set("" value("PID")); };
> Aug  2 11:11:11 test test: test

Signed-off-by: László Várady <laszlo.varady@balabit.com>