logproto: buffer size from state
This description ignores the char encoding convertion, as that is not effected by
the bug, as that has buffer size reallocation built into, while without
conversion the log-msg-size allocated.

when configured the log-msg-size does correlate to an internal buffer size,
stored in LogProtoBufferedServer. The LogProtoBufferedServer actually just gets
the message size from log-msg-size if it does not have state stored in persist
before.

In case of LogProtoTextServer (is a LogProtoBufferServer) the message is split
either by delimiter ('\n') or via the maximum message size (log-msg-size).
Because the LogProtoBufferedServer only checks once the log-msg-size and stores
that value into the persist, when the log-msg-size is change (stop->config,
change->start), the new value is no longer used for its internal buffer.
The value is still correctly used as max message value in case of
LogProtoTextServer for splitting up the messages.

The LogProtoTextServer

if there is space in its buffer ask transport to fill up the empty space
tries to find the message end: either via separator ('\n'), or simply because the
 max message size has been reached (log-msg-size)

At configuration change, the log-msg-size could change in tree possible way:
(a) do not change
(b) decrease
(c) increase

The scope is only the LogProtoBufferServer's buffer and its buffer_size
(a) This is the same as starting freshly without a persist state. No issue here.
(b) Here a little larger bufffer is used, that is really needed. Also the first
    message respects the previose configuration log-msg-size if there is a
    separator in it ('\n').
(c) When the log message size is smaller than the previose log-msg-size,
    the message is delivered just fine, no issue.
But when the message is larger then the previose log-msg-size, the buffer cannot
hold the whole message, therefore there is no separator ('\n'). Splitting up the
message a desired way with separator is not possible.

The next line of defense is the new/current log-msg-size that splits up too long
messages. But this can never be a too long message as the message max length
(according to syslog-ng knowladge) is previose log-msg-size what is smaller than
the next/current log-msg-size. This logic could not detect either a new message.
There is only this two logic to find message in the buffer, so syslog-ng cannot
find any message and stuck in this state foreever.

The proposed solution is to let syslog-ng choose the bigger value for
LogProtoBufferServer's buffer either the one in its persist state or the one
configured via (log-msg-size).

Signed-off-by: Kokan <kokaipeter@gmail.com>