affile: detect EOF as soon as possible
The EOF detection has been moved from the timer to update_watches(),
so the end of the file can be detected right after reading from it.

The EOF detection will be used to implement multi-line-timeout().
Currently it is only used by the wildcard file source.
update_watches() is also called periodically (follow-freq), so this
change does not alter how EOF is signaled.

Signed-off-by: László Várady <laszlo.varady@balabit.com>