affile: add PollMultilineFileChanges
PollMultilineFileChanges is responsible for polling and handling
multi-line file changes.
Since the multi-line file source detects the end of a message after
finding the beginning of the subsequent message, additional actions
are required on certain events:

- when the file is moved/restarted, the last (potentially partial)
message should be flushed
- after waiting a preconfigured time (multi-line-timeout()) without
reading new data, the last message should also be flushed.

Signed-off-by: László Várady <laszlo.varady@balabit.com>