add-contextual-data: remove core->"plugin" dependency
The intent seemingly was to make the add-contextual-data core independent
of the specifics of various selector "plugins", e.g. the template and
the filter based selectors.

This was violated as the constructors for these plugins were part of the
core. With the latest cleanups these have become one-liners, so move
them to the grammar file and remove this dependency from add-contextual-data.c

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>