kafka: fix a per-message memory leak in kafka destination
In case the key() option is not used, we are allocating a g_malloc(1) memory
for every message that never gets freed. Although in small bites, but we
can eventually leak a large amount of memory.

Thanks to the efforts of @jbfuzier we got enough information to
troubleshoot the issue.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>