logproto-buffered-server: fix an issue with bookmark/persist handling
LogProtoBufferedServer is designed to work without persist state if the
allocation fails. (persist-state will schedule a mainloop exit in this
case, but LPBServer will keep runing until we shut down gracefully)

The statless mode is handled internally by returning a dummy state in
case of allocation failure. There are safeguards in _get_state, and
_put_state to not actually try to "persist_state_map_enty" this value.

However in Bookmark handling we only pass the persist_handle to the
callback function, not the whole LogProtoBufferedServer, and we can not
use our safeguards to prevent 'persist_state_map_entry' that invalid
value. This patch separates that case by using a different callback
function.

Signed-off-by: Laszlo Szemere <laszlo.szemere@oneidentity.com>