logproto: add a LogProtoBufferedServer method to flush buffers
The log_proto_buffered_server_cue_flush() method makes the buffered server
flush its buffer during the subsequent fetch() call.

The partial content in the buffer will be stored in a LogMessage instance,
which will be sent through the pipeline.

In the older versions of syslog-ng (PE 6), the fetch() method was called
directly in flush(), but this would execute the whole pipeline (filters,
parsers, etc.) in the thread of the caller (main thread).

log_proto_buffered_server_cue_flush() can be used together with a timer,
so partial messages can be flushed after a timeout (if no new data is
received).

Signed-off-by: László Várady <laszlo.varady@protonmail.com>