afpipe: add create-dirs option to pipe source
Pipe source creates the FIFO where it will read,
thus it should be able to create the path as well.

If we don't release this option and thus the feature,
then syslog-ng can FAIL to start up in cases,
with a configured pipe source without the
existing directories in the path.

Please note that previously file/pipe sources silently
did create the directories recursively,
and it was impossible to disable directory creation.
What I mean under silently, is that the `create_dirs` option
left uninitialized (-1), and FileOpener considered this as True.

About file source: it does not need the option create-dirs(),
as it will poll the path, and it shouldn't create the directories
since that is the job of the destination (the one who writes into the file).

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>