logmsg: add default pri value to log_msg_init
In case it's left uninitialized, the message could
seem as having a pri LOG_KERN | LOG_EMERG.

This is only interesting, if a user would like to
create a log message, but without initial parsing,
thus log_msg_new() is not suitable, as it needs a format_handler, e.g. syslog .
Currently this is done by log_msg_new_empty(), which is used by some drivers.

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>