syslogformat: fuzz input with libfuzzer
The libfuzzer is a code coverage based fuzzer provided by LLVM[1]. The aim of
this fuzzing target introduced to the syslogformat to mitigate ill-formed inputs
breaking the application, causing crash, buffer overflow, null pointer
dereference. In order to detect such issues the fuzzer can and should be
combined with sanitizers such as ASAN, MSAN, UBSAN[2]. By using the existing
`-DSANITIZER={}` option to specify different kind of sanitizers (multiple is
possible), the sanitizers listed with that option is going to be combined with
the fuzzer. The libfuzzer could use `corpus`-es as initial inputs, patterns or
dictionary to compose new inptus based on coverage and randomness. Due to the
random execution it is not a set of tests that should be executed for each code
change, rather a test that should be exectued continuously. The target is not
part of the default UT test set, it must be executed manually as of now.

In order to run the fuzzer one could simply call the executable and wait as long
as it finds some crash - at which point it is going to stop, or a desired amount
of time and stop the executable, as otherwise it runs indefenetly by default.

The issues found in the next patchset is found by the following specialization of
the fuzzer:
```
./modules/syslogformat/tests/test_fuzzing_syslogformat ../modules/syslogformat/tests/corpus/ -print_pcs=1 -detect_leaks=1 -print_final_stats=1 -print_corpus_stats=1 -use_value_profile=1  -max_len=256
```

The `-use_value_profile=1` option helped to find most of the issues, despite its
being currently an experimental feature.

[1] http://llvm.org/docs/LibFuzzer.html
[2] http://llvm.org/docs/LibFuzzer.html#fuzzer-usage

Signed-off-by: Kokan <kokaipeter@gmail.com>