plugin: allow multiple discoveries of plugin candidates
Sometimes the `module-path` define changes within the configuration,
and it makes sense to re-run the discovery in case that happens. So
allow the higher layer (cfg.c in this case) to decide if discovery
needs to rerun, so plugin_discover_candidate_modules() will now drop
the old candidate list and blindly rerun the discovery in case it
gets called.

This patch will only change the plugin logic to handle multiple invocations
but does not change the higher layers to actually invoke
plugin_discover_candidate_modules() multiple times per configuration.

The only functional change is that previously SIGHUPs wouldn't rescan the
modules, with this patch, it would.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>