affile: LogProtoFileWriter should return LPS_PARTIAL in case the partial buffer is not completely emptied
When a writev() issues a short write (e.g.  less bytes written than
requested), we are moving the iovec into a local buffer (that probably
shouldn't be moved, but anyway).

LogProtoFileWriter was returning LPS_SUCCESS while this buffer is being
emptied.

1) log_proto_client_post(): we returned LPS_SUCCESS && consumed==FALSE

2) log_proto_client_flush(): we returned LPS_SUCCESS without a means to
indicate when we are finished.

This didn't cause issues when LogWriter is using the LogProtoClient instance
as that would simply try to push messages to the proto as they are flowing
in.  Whenever we exit, it just tries a single flush() and then gives up.

However, this prevents me to write a unit test around the case where the
transport is generating short writes, as we never know when we get into a
consistent state with the buffer empty.

LPS_PARTIAL is there for this case and is being used in LogProtoTextClient
for the same case, so let's return LPS_PARTIAL here too.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>