cfg-tree: destinations must not be linked as chain.
We want to leave pipe-next available to use for destinations. Prior
this patch, pipe-next was used by the config graph to pass messages
forward in a sequence layout. But pipe-next was overridden by network
destination (pointing to LogWriter), hence disjointing the config graph.

This patch links destinations in T form, instead of single link.

```
* (endpoint of sequence)
|
V
* (multiplexer) -(next-hop)-> destination - (pipe-next*) -> logwriter
|
(pipe-next)
|
V
* (rest of the sequence)
```

That way destinations are free to use the pipe-next*

Fixes: https://github.com/syslog-ng/syslog-ng/issues/2820

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>