python-dest: refactor use connect fuction instead of thread_init
The connect-disconnect pair is clearer than thread_init-disconnect.

One might think that this is a functional change, because the connect
method might be delayed until the first message. Apparently this is
not the case: threaded destination driver also start with a connect
without any message (this is not straightforward from the code).

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>