travis: install pre-commit-hooks only when needed
This time of writing, building arm in travis broke because it cannot
install pre-commit-hooks. Only the arm build is affected.
Unfortunately I do not know the reason. The rollback to previous
version of pre-commit-hooks did not help. And I could not reproduce
the problem myself other than travis. So I just made sure
pre-commit-hooks is installed only in those cases when needed.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>