scl: make a missing optional modules fatal
In case elasticsearch-http()/slack()/telegram() was used without the
http() module being present gave us an error about a missing
elasticsearch-http() driver, which was actually present as part of
the SCL, but its processing was skipped because of a @requires in front.

This patch pushes @requires into the body of the block, thus it will
only be processed whenever these driver are used in the user's
configuration.

And it also makes that requires fatal together with a nice message
with the reason of the error.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>