lex: pass all characters from the config to the parser
Syslog-ng has a (not well documented) feature to accept strings
without quotation marks. One of the key benefit of this feature
is that we do not need to put identifiers (i.e.: block names)
into quotation marks. Which makes the config much more readable.

We achieve this by an internal lookup. After Flex successfully
identified a "word" (for details, please investigate the rule
inside the lexer file), we will search for it in the list of
known keywords. If there is no match, we will handle that token
as a string.

Flex - by design - ignores inputs without matching patterns.
(Writes them on standard error, and continues parsing.)

As a sum of this two mechanism, if a character left out from the
definition of a "word", and there were no other rules to pick it
up, than we simply ignored it.
Saying it differently: Strings could lose special leading or
trailing characters without quotation marks.

As an example, the triggering issue was a problem with the prefix
option of the XML parser: "xml(prefix(.SDATA.))"
In this case the XML parser was initialized successfully, there
were no error messages from Syslog-ng. However the final value of
the prefix string inside XML parser was: "SDATA".

Signed-off-by: Laszlo Szemere <laszlo.szemere@oneidentity.com>