transport: fix infinite loop when TLS connection is broken
syslog-ng does not call SSL_connect() or SSL_do_handshake(), so the first
SSL_read() or SSL_write() call is responsible for negotiating the TLS
session. Nevertheless, the initial IO cond should be 0, because the
LogProto layer should choose the initial IO direction.

In case TLS renegotiation is required, OpenSSL returns an error
(SSL_ERROR_WANT_READ or SSL_ERROR_WANT_WRITE). In this case, the same
read/write function must be invoked after a IO readiness notification.

With all that being said, the initial G_IO_IN | G_IO_OUT transport-level
IO-cond is not necessary and even harmful:

If the connection is lost in a TLS destination before sending the first
message, an infinite IO loop starts in the main thread. This is because
ivykis reports readiness for reading, but we won't read anything as the
LogProto instance is not in the correct state.

Detecting the breakdown of the connection was not possible because we
would have to read from the socket first, and only then could we notice
about the closed connection.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>