MergeYm: skip `rewrite-expr-grammar.ym`
Currently we are not collecting the `rewrite` options, so it is not
needed.

However, the bigger reason to exclude it, is because we shorten the
database build time from ~30 sec to ~2 sec. This is because of the
unique syntax of the rewrite rules, namely these lines:

```
start
    : rewrite_expr_list log_last_junction { ... }
    ;

rewrite_expr_list
    ...
    | log_junction semicolons rewrite_expr_list { ... }
    ;
```

`log_junction` and `last_log_junction` builds the whole config tree
again, with ~13 million paths. (Just for comparison all the other
grammar files produce ~30.000 paths)

We will tackle this problem later, when we decide to process rewrite
rules.

Signed-off-by: Attila Szakacs <attila.szakacs@balabit.com>