python-persist: kick-off code to increase persist size
This time of writing, persist file can be increased in ~pagesize
bytes. A proper implementation would allow to increase arbitrary
size. But that would need to be implemented in lib/persist.

This patch only allows to store pagesize/2 at a time, which sounds
reasonable limitation. Persist should not be used as a large storage.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>