syslogformat: add corpuses for fuzzing test
The libfuzzer could work with or without any predefined patterns. By providing
patterns it faster to find different execution paths, always a good idea. There
are two possible way to provide samples (potentially valid inputs) or provide
dictionary containing atomic component that possible to build up structured
inputs.

For the syslogformat the corpus was choosen, because the input does not depend on
having the structural construtions like opening closing bracket. It is more loos
and relaxed, contains some heuristic based on seen inputs.

The libfuzzer uses the corpus directory as both an input and output, the newly
found inputs that provides unseen path it saves later.
The `modules/syslogformat/tests/corpus` directory is by default ignored via
`.gitignore` in order not to save all of the generated new corpuses. It is still
possible to force git to add new corpuses.

Signed-off-by: Kokan <kokaipeter@gmail.com>