loggen: improve address resolution
This is a fix for the same problem that #2933 resolves in syslog-ng.

"When AI_ADDRCONFIG is used, then IPv4 addresses are returned in the list
pointed to by res only if the local system has at least one IPv4 address
configured,  and  IPv6 addresses  are  returned only if the local system
has at least one IPv6 address configured. The loopback address is not
considered for this case as valid as a configured address.

This flag is useful  on,  for  example,  IPv4-only systems, to ensure
that getaddrinfo() does not return IPv6 socket addresses that would
always fail in connect(2)."

Retrying an unsuccessful resolution without AI_ADDRCONFIG is necessary
because ::1 or 127.0.0.1 might not be resolved when the AI_ADDRCONFIG
flag is set.

AI_V4MAPPED has also been added to support IPv4-mapped IPv6 addresses
when no matching IPv6 addresses could be found.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>