cfg-grammar: separate reference-, inline-template rules
The current rule handles both `template(foo)` and `template("foo")` options, in
both case implying the quoted/string version. The rule first looks for existing
template definitions with the exact same identifier or string if there is one,
that template definition will be used(reference mode), otherwise it tries to
parse the identifier or string as a template format(inline mode). In any other
cases yields invalid configuration.

The following example might not work as one would expect:
```
@version: 3.25

log {
  destination {
    file("/dev/stdout" template(foo));
  };
};

template foo "bar";
```

The "foo" text is used as a template, and not the - most likely - intended "bar".

This change modifies the `template()` behavior based on what is provided as
argument:
* quoted string (either `'` or `"`) will be interpreted as inline template
* non-quoted argument is handled as reference template

Implication of the above change for the user:

* Allows user to write `template("foo")` that computes as `foo`
  (even if `foo` is already defined as a template)
* Yield configuration error for `template(foo)` if `foo` is not defined as
  template, or defined after the current use.
  (As of now in this case there is no error, but simply falls back to the previose
   use case.)

This is a breaking change, but in favor of the user intent.

Signed-off-by: Kokan <kokaipeter@gmail.com>